// ui/src/features/services/api/services-list.generated.ts
type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
// ...
export type ServicesListQueryVariables = Exact<{
  fromNanos: string
  toNanos: string
}>

export type ServicesListQuery = {
  readonly serviceList: ReadonlyArray<{
    readonly name: string
    readonly lastSeenNanos: string
    readonly spanCount: string
    readonly errorCount: string
    readonly p95Ms: number | null
  }>
  readonly serviceCatalog: ReadonlyArray<{
    readonly name: string
    readonly serviceVersion: string | null
    readonly serviceNamespace: string | null
    readonly deploymentEnvironment: string | null
    readonly telemetrySdkLanguage: string | null
    readonly telemetrySdkName: string | null
    readonly telemetrySdkVersion: string | null
    readonly lastSeenNanos: string
    readonly instanceCount: string
  }>
}

export const ServicesListDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "query",
      name: { kind: "Name", value: "ServicesList" },
      variableDefinitions: [
        {
          kind: "VariableDefinition",
          variable: {
            kind: "Variable",
            name: { kind: "Name", value: "fromNanos" },
          },
          type: {
            kind: "NonNullType",
            type: {
              kind: "NamedType",
              name: { kind: "Name", value: "String" },
            },
          },
        },
        {
          kind: "VariableDefinition",
          variable: {
            kind: "Variable",
            name: { kind: "Name", value: "toNanos" },
          },
          type: {
            kind: "NonNullType",
            type: {
              kind: "NamedType",
              name: { kind: "Name", value: "String" },
            },
          },
        },
      ],
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            name: { kind: "Name", value: "serviceList" },
            arguments: [
              {
                kind: "Argument",
                name: { kind: "Name", value: "fromNanos" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "fromNanos" },
                },
              },
              {
                kind: "Argument",
                name: { kind: "Name", value: "toNanos" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "toNanos" },
                },
              },
            ],
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                { kind: "Field", name: { kind: "Name", value: "name" } },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "lastSeenNanos" },
                },
                { kind: "Field", name: { kind: "Name", value: "spanCount" } },
                { kind: "Field", name: { kind: "Name", value: "errorCount" } },
                { kind: "Field", name: { kind: "Name", value: "p95Ms" } },
              ],
            },
          },
          {
            kind: "Field",
            name: { kind: "Name", value: "serviceCatalog" },
            arguments: [
              {
                kind: "Argument",
                name: { kind: "Name", value: "fromNanos" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "fromNanos" },
                },
              },
              {
                kind: "Argument",
                name: { kind: "Name", value: "toNanos" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "toNanos" },
                },
              },
            ],
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                { kind: "Field", name: { kind: "Name", value: "name" } },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "serviceVersion" },
                },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "serviceNamespace" },
                },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "deploymentEnvironment" },
                },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "telemetrySdkLanguage" },
                },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "telemetrySdkName" },
                },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "telemetrySdkVersion" },
                },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "lastSeenNanos" },
                },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "instanceCount" },
                },
              ],
            },
          },
        ],
      },
    },
  ],
} as unknown as DocumentNode<ServicesListQuery, ServicesListQueryVariables>
// ...
export const ServicesListQuerySchema: z.ZodType<ServicesListQuery> = z.object({
  serviceList: z.array(
    z.object({
      name: z.string(),
      lastSeenNanos: z.string(),
      spanCount: z.string(),
      errorCount: z.string(),
      p95Ms: z.number().nullable(),
    })
  ),
  serviceCatalog: z.array(
    z.object({
      name: z.string(),
      serviceVersion: z.string().nullable(),
      serviceNamespace: z.string().nullable(),
      deploymentEnvironment: z.string().nullable(),
      telemetrySdkLanguage: z.string().nullable(),
      telemetrySdkName: z.string().nullable(),
      telemetrySdkVersion: z.string().nullable(),
      lastSeenNanos: z.string(),
      instanceCount: z.string(),
    })
  ),
})